# src/cisco.com/son/apphcd/api/v1/appmanager/appmanager.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "appmanager.proto",
    "version": "version not set"
  },
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/api/v1/apps": {
      "get": {
        "summary": "GetApps shows appropriate information about running application",
        "operationId": "GetApps",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/appmanagerResponse"
            }
          }
        },
        "parameters": [
          {
            "name": "name",
            "description": "Application name.",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "version",
            "description": "Application version.\nIf omitted, controller will provide information about all running application instances.",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "cycle",
            "description": "Application type.",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "root_group_id",
            "description": "Root Group ID.",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "group_ids",
            "description": "A list of group IDs.",
            "in": "query",
            "required": false,
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          {
            "name": "verbose",
            "description": "Printout a lot of information.",
            "in": "query",
            "required": false,
            "type": "boolean",
            "format": "boolean"
          }
        ],
        "tags": [
          "AppManager"
        ]
      },
      "delete": {
        "summary": "DeleteApps deletes running applications and related workloads.\nIf \"purge\" flag is set True, the applications will be also removed from the catalog",
        "operationId": "DeleteApps",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/appmanagerResponse"
            }
          }
        },
        "parameters": [
          {
            "name": "purge",
            "description": "Indicates whether all the apps should be also removed from catalog.",
            "in": "query",
            "required": false,
            "type": "boolean",
            "format": "boolean"
          }
        ],
        "tags": [
          "AppManager"
        ]
      },
      "post": {
        "summary": "CreateApp method creates a new application.\nIt will create a new catalog entry and install the new application from the catalog.\nIf appropriate application instance with the same name and version already running,\nthe instance won't be overwritten and no error will be returned.The request will continue\nto the next instance.If all requested instances already running, the status \"UNCHANGED\" will be\nreturned to the caller. A list of created instances will be returned as part of the response body",
        "operationId": "CreateApp",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/appmanagerResponse"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/appmanagerCreateAppRequest"
            }
          }
        ],
        "tags": [
          "AppManager"
        ]
      }
    },
    "/api/v1/apps/activation": {
      "post": {
        "summary": "EnableDisableApp disables or enables an application",
        "operationId": "EnableDisableApp",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/appmanagerResponse"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/appmanagerEnableDisableAppRequest"
            }
          }
        ],
        "tags": [
          "AppManager"
        ]
      }
    },
    "/api/v1/apps/metadata/{app_name}": {
      "delete": {
        "summary": "DeleteAppMetadata deletes metadata for a particular application instance",
        "operationId": "DeleteAppMetadata",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/appmanagerResponse"
            }
          }
        },
        "parameters": [
          {
            "name": "app_name",
            "description": "Application name",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/appmanagerDeleteAppMetadataRequest"
            }
          }
        ],
        "tags": [
          "AppManager"
        ]
      }
    },
    "/api/v1/apps/{name}": {
      "delete": {
        "summary": "DeleteApp deletes instances of a particular running application.\nIt's possible to customize the request by providing appropriate body",
        "operationId": "DeleteApp",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/appmanagerResponse"
            }
          }
        },
        "parameters": [
          {
            "name": "name",
            "description": "Application name.",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/appmanagerDeleteAppRequest"
            }
          }
        ],
        "tags": [
          "AppManager"
        ]
      },
      "put": {
        "summary": "UpgradeApp creates new, upgrades or downgrades existing application.\nApplication will be recreated if the version in request is equal to the version of running application.\nIn this case the catalog artifact will be replaced with the new one the running application will be replaced\nwith recent application from the catalog.\nApplication will be upgraded if the version in request is higher than version of running application.In this case\na new catalog artifact will be created and the running application will be upgraded from the catalog.\nApplication will be downgraded if the version in request is lower than version of running application. In this case\na new catalog artifact will be created and the running application will be downgraded from the catalog",
        "operationId": "UpgradeApp",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/appmanagerResponse"
            }
          }
        },
        "parameters": [
          {
            "name": "name",
            "description": "Application name.",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/appmanagerUpgradeAppRequest"
            }
          }
        ],
        "tags": [
          "AppManager"
        ]
      },
      "patch": {
        "summary": "UpdateApp updates existing application instances. The new application instance will be based on configuration\nof existing instance hence a minimal request can be sent to controller. If there is no application instance\nrunning, the method will return \"UNCHANGED\" status and message \"Nothing to update/upgrade\".\nThis method always recreates existing application instance",
        "operationId": "UpdateApp",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/appmanagerResponse"
            }
          }
        },
        "parameters": [
          {
            "name": "name",
            "description": "Application name.",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/appmanagerUpdateAppRequest"
            }
          }
        ],
        "tags": [
          "AppManager"
        ]
      }
    }
  },
  "definitions": {
    "CyclePeriodicReqAttrSched": {
      "type": "object",
      "properties": {
        "Saturday": {
          "type": "boolean",
          "format": "boolean"
        },
        "Sunday": {
          "type": "boolean",
          "format": "boolean"
        },
        "Monday": {
          "type": "boolean",
          "format": "boolean"
        },
        "Tuesday": {
          "type": "boolean",
          "format": "boolean"
        },
        "Wednesday": {
          "type": "boolean",
          "format": "boolean"
        },
        "Thursday": {
          "type": "boolean",
          "format": "boolean"
        },
        "Friday": {
          "type": "boolean",
          "format": "boolean"
        }
      }
    },
...
    "appmanagerAppStateAfterDeployment": {
      "type": "string",
      "enum": [
        "disabled",
        "enabled"
      ],
      "default": "disabled",
      "description": "- disabled: Enable application after deployment\n - enabled: Disable application after deployment",
      "title": "Defines application instance state"
    },
    "appmanagerCreateAppRequest": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "title": "Application name"
        },
        "app_state": {
          "$ref": "#/definitions/appmanagerAppStateAfterDeployment",
          "title": "Enable or disable the application after deployment. By default application disabled"
        },
        "version": {
          "type": "string",
          "title": "Upgrade existing application instance to the version exposed by the attribute"
        },
        "from_catalog": {
          "type": "boolean",
          "format": "boolean",
          "title": "Don't create a new metadata but use existing one. Installs directly from catalog"
        },
        "description": {
          "type": "string",
          "description": "Application description."
        },
        "cycle": {
          "type": "string",
          "title": "Application type. Supported types:\nperiodic - application of type \"CronJob\" will be created, will be scheduled according to\nthe CyclePeriodic configuration.\ndaemon - application of type \"Deployment\" will be created, will be running continuously\nrun_once - application of type \"Job\", will run once until success"
        },
        "cycle_periodic_attr": {
          "$ref": "#/definitions/appmanagerCyclePeriodicReqAttr",
          "title": "Cycle Periodic attributes"
        },
        "root_group_id": {
          "type": "string",
          "title": "Root Group ID"
        },
        "group_ids": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "title": "A list of group ID"
        },
        "app_configs": {
          "type": "object",
          "additionalProperties": {
            "type": "string"
          },
          "title": "Application configurations"
        },
        "env_vars": {
          "type": "object",
          "additionalProperties": {
            "type": "string"
          },
          "title": "Application environment variables"
        },
        "secrets": {
          "type": "object",
          "additionalProperties": {
            "type": "string"
          },
          "title": "Secret objects"
        },
        "labels": {
          "type": "object",
          "additionalProperties": {
            "type": "string"
          },
          "title": "Labels"
        },
        "annotations": {
          "type": "object",
          "additionalProperties": {
            "type": "string"
          },
          "title": "Annotations"
        },
        "shared_storage": {
          "type": "integer",
          "format": "int64",
          "title": "Dependencies\nrepeated Dependency deps = 12;\nShared storage size"
        },
        "spec": {
          "$ref": "#/definitions/appmanagerSpec",
          "title": "Instance specifications"
        }
      },
      "title": "CreateAppRequest holds appropriate properties required for a particular request\nrelated to a new application instance creation"
    },
    "appmanagerCyclePeriodicReqAttr": {
      "type": "object",
      "properties": {
        "working_days": {
          "$ref": "#/definitions/CyclePeriodicReqAttrSched"
        },
        "min_start_hour": {
          "type": "integer",
          "format": "int64"
        },
        "max_start_hour": {
          "type": "integer",
          "format": "int64"
        },
        "interval_min": {
          "type": "integer",
          "format": "int64"
        }
      },
      "title": "CyclePeriodicReqAttr message holds information for an application of type periodic"
    },
    "appmanagerDeleteAppMetadataRequest": {
      "type": "object",
      "properties": {
        "app_name": {
          "type": "string",
          "title": "Application name"
        },
        "version": {
          "type": "string",
          "title": "Deletes only a particular version of application instance"
        },
        "root_group_id": {
          "type": "string",
          "title": "Root Group ID"
        },
        "group_ids": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "title": "A list of group IDs"
        }
      },
      "title": "DeleteAppMetadataRequest holds attributes required for deleting\nmetadata for appropriate application and related instances"
    },
    "appmanagerDeleteAppRequest": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "description": "Application name."
        },
        "version": {
          "type": "string",
          "description": "Application version."
        },
        "root_group_id": {
          "type": "string",
          "title": "Root Group ID"
        },
        "group_ids": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "title": "A list of group IDs"
        },
        "purge": {
          "type": "boolean",
          "format": "boolean",
          "title": "Indicates whether the app should be also removed from catalog"
        }
      },
      "title": "DeleteAppRequest holds attributes required for deleting\nappropriate application and its related instances"
    },
    "appmanagerEnableDisableAppRequest": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "title": "Application name"
        },
        "version": {
          "type": "string",
          "title": "Application version"
        },
        "root_group_id": {
          "type": "string",
          "title": "Root Group ID"
        },
        "group_ids": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "title": "A list of group IDs"
        },
        "disable": {
          "type": "boolean",
          "format": "boolean",
          "title": "Enable or disable application instances"
        }
      },
      "title": "EnableDisableAppRequest holds attributes required for disabling or enabling appropriate application instance"
    },
    "appmanagerResponse": {
      "type": "object",
      "properties": {
        "timestamp": {
          "type": "string",
          "format": "date-time"
        },
        "status": {
          "$ref": "#/definitions/appmanagerStatus"
        },
        "message": {
          "type": "string"
        },
        "body": {
          "$ref": "#/definitions/protobufAny"
        }
      },
      "title": "Response holds information related to a response message that is sent on appropriate request"
    },
    "appmanagerSpec": {
      "type": "object",
      "properties": {
        "image": {
          "$ref": "#/definitions/SpecImage"
        },
        "ports": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/appmanagerSpecPort"
          }
        },
        "resources": {
          "$ref": "#/definitions/appmanagerSpecResources"
        }
      },
      "title": "Specification message"
    },
...
    "appmanagerStatus": {
      "type": "string",
      "enum": [
        "SUCCESS",
        "ERROR",
        "NOT_FOUND",
        "UNCHANGED"
      ],
      "default": "SUCCESS",
      "description": "Status represents operation status.\n\n - SUCCESS: Operation was successful\n - ERROR: Operation failed\n - NOT_FOUND: Resource not found\n - UNCHANGED: Operation had no results (e.g. upgrade identical, rollback to same, delete non-existent)"
    },
...
  }
}